+++ lib/src/models/call_settings_model.dart
@@ -2,5 +2,5 @@ class CallSettingsModel {
   final Duration duration;
   final bool isVibrationOn;
 
-  CallSettingsModel(this.duration, this.isVibrationOn);
+  const CallSettingsModel(this.duration, this.isVibrationOn);
 }