+++ lib/src/utilities/size_config.dart
@@ -46,20 +46,3 @@ class VerticalSpacing extends StatelessWidget {
     );
   }
 }
-
-// For add free space horizontally
-class HorizontalSpacing extends StatelessWidget {
-  const HorizontalSpacing({
-    Key? key,
-    this.of = 20,
-  }) : super(key: key);
-
-  final double of;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: getProportionateScreenWidth(of),
-    );
-  }
-}