+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 1.0.1
+* Fixes close button bug on launch
+* Updates read me
+* Refactors widgets
+
+## 1.0.1
 * Make video path optional
 * Fix Bugs
 * Refactor widgets